GameClient and GameServer are now singletons.
Added a ton of new classes.

The server will start up a simulation and listen for connections. A client will now connect to the localhost and request joining. The server will add the requested client to the simulation and notify all players of the new client.

Also renamed JoinMesssage to JoinMessage.

+++ Shared/src/edu/spaced/net/listener/ChatListener.java
@@ -1,5 +1,5 @@
 package edu.spaced.net.listener;
 
 public interface ChatListener {
-	void chatMessage(String text, long from);
+	void chatMessage(String text, int from);
 }
